requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
alabaster==0.7.13
antiword==0.1.0
anyio==4.0.0
APScheduler==3.9.1.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Automat==22.10.0
Babel==2.14.0
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.1.0
blinker==1.6.3
blis==0.7.11
boto3==1.28.71
botocore==1.31.71
CacheControl==0.12.14
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
comm==0.2.0
comtypes==1.2.0
confection==0.1.3
ConfigArgParse==1.7
confluent-kafka==2.3.0
constantly==15.1.0
cryptography==41.0.5
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.6
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-db-connection-pool==1.2.4
django-db-geventpool==4.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
docx==0.2.4
docx2txt==0.8
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fakeredis==2.20.0
fastjsonschema==2.18.1
fbmessenger==4.3.1
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
gevent==23.9.1
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.18.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
incremental==22.10.0
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.19
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==7.4.9
jupyter_core==5.4.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
keras==2.12.0
kiwisolver==1.4.5
klein==23.5.0
langcodes==3.3.0
langid==1.1.6
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
locket==1.0.0
lxml==4.9.3
magic-filter==1.0.12
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mattermostwrapper==2.2
mistune==3.0.2
ml-dtypes==0.3.1
mpmath==1.3.0
msgpack==1.0.7
multidict==5.2.0
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nbsphinx==0.9.3
neat-python==0.92
nest-asyncio==1.5.8
networkx==2.6.3
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
olefile==0.47
opencv-python==4.8.1.78
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pamqp==3.2.1
pandoc==2.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pathlib==1.0.1
pathy==0.10.3
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
plumbum==1.8.2
ply==3.11
pooch==1.8.0
portalocker==2.8.2
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.3
psutil==5.9.7
psycogreen==1.0.2
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
pygame==2.5.2
Pygments==2.16.1
PyJWT==2.8.0
pykwalify==1.6.0
pymatreader==0.0.32
pymongo==4.3.3
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyparsing==3.1.1
pyreadline3==3.4.1
pyrsistent==0.20.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-docx==1.1.0
python-engineio==4.8.0
python-json-logger==2.0.7
python-socketio==5.10.0
python-telegram-bot==20.6
pytz==2022.7.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==24.0.1
questionary==1.10.0
randomname==0.1.5
redis==4.6.0
referencing==0.32.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rocketchat-API==1.30.0
rpds-py==0.15.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.11.3
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.14.0
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.23.0
slackclient==2.9.4
smart-open==6.4.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==1.4.49
sqlparams==6.0.1
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
structlog==23.2.0
structlog-sentry==2.0.3
sympy==1.12
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminado==0.18.0
terminaltables==3.1.10
thinc==8.1.12
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.26.0
Tubes==0.2.1
twilio==8.2.2
Twisted==23.8.0
twisted-iocpsupport==1.0.4
typer==0.7.0
types-python-dateutil==2.8.19.14
typing==3.7.4.3
typing-utils==0.1.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.7
wasabi==0.10.1
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.1
Whoosh==2.7.4
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.1.1
xmltodict==0.13.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.1
